Add style to index.

+++ rest-api/php/php-client-example/index.php
@@ -19,7 +19,7 @@ if (file_exists(dirname(__FILE__)  . '/local_config.php')) {
 }
 
 // Make sure you keep your PK12 key file in a secure location, and isn't readable by others.
-const KEY_FILE = $apiConfig['xtuple']['oauth2_pk12_filename'];
+$key_file = $apiConfig['xtuple']['oauth2_pk12_filename'];
 
 $client = new Google_Client();
 
@@ -30,14 +30,14 @@ $client::$io->setOptions(array (CURLOPT_SSL_VERIFYPEER => !$apiConfig['xtuple'][
 // Set your cached access token. Remember to replace $_SESSION with a
 // real database or memcached.
 session_start();
-session_unset();
-if (isset($_SESSION['token'])) {
-  $client->setAccessToken($_SESSION['token']);
+//session_unset();
+if (isset($_SESSION['access_token'])) {
+  $client->setAccessToken($_SESSION['access_token']);
 }
 
 // Load the key in PKCS 12 format (you need to download this from the Mobile
 // Client Admin Interface when the OAuth 2.0 service account client was created.
-$key = file_get_contents(KEY_FILE);
+$key = file_get_contents($key_file);
 $oauth = new Google_AssertionCredentials(
   $apiConfig['oauth2_client_id'],
   array($apiConfig['xtuple']['oauth2_scope']),
@@ -53,6 +53,14 @@ $client->setAssertionCredentials($oauth);
 
 // Create a new service client.
 $service = new Rescued_ApiService($client, $apiConfig['xtuple']);
+
+echo "<html>
+        <head>
+          <link rel='stylesheet' href='style.css' />
+        </head>
+        <body class='html not-front one-sidebar sidebar-first'>
+        <div class='column sidebar' id='sidebar-first'><p>&nbsp;</p></div>
+        <div class='column' id='content'>";
 
 // Below are some very basic DELETE, GET, PATCH and POST request examples.
 // These examples are intended to show you how to interact with xTuple's
@@ -77,31 +85,69 @@ $service = new Rescued_ApiService($client, $apiConfig['xtuple']);
 if (!isset($_GET['method']) && !isset($_GET['id'])) {
   $key = get_resource_key_field($service, $apiConfig, 'Contact');
   $added =  false;
+  $output = '';
 
   // Make the GET request.
   $result = $service->Contact->list();
-
-  echo "<h2>List of Contacts:</h2>";
-    echo "<pre>";
-   // print_r($result);
-    echo "</pre>";
-
   foreach ($result['data'] as $value) {
-    echo "<h3>Contact: ";
-    echo "<a href='index.php?method=GET&id=" . $value[$key] . "'>" . $value[$key] . "</a>";
-    echo "</h2>";
-    echo "<p>";
-    echo $value['firstName'] . " " . $value['lastName'];
-    echo "</p>";
+    $output .= "<tr>
+            <td>";
+    $output .= "<a href='index.php?method=GET&id=" . $value[$key] . "'>" . $value[$key] . "</a>";
+    $output .= "  </td>";
+    $output .= "  <td>";
+    $output .= $value['firstName'];
+    $output .= "  </td>";
+    $output .= "  <td>";
+    $output .= $value['lastName'];
+    $output .= "  </td>";
+    $output .= "  <td>";
+    $output .= $value['phone'];
+    $output .= "  </td>";
+    $output .= "  <td>";
+    $output .= $value['primaryEmail'];
+    $output .= "  </td>";
+    $output .= "  <td>";
+    $output .= "<a href='index.php?method=GET&id=" . $value[$key] . "'> View </a>";
+    $output .= " - <a href='index.php?method=PATCH&id=" . $value[$key] . "'> Add Comment </a>";
+    if ($value[$key] === '42') {
+      $output .= " - <a href='index.php?method=DELETE&id=" . $value[$key] . "'> Delete </a>";
+    }
+    $output .= "  </td>
+          </tr>";
 
     if ($value[$key] === '42') {
       $added = true;
     }
   }
 
+  echo "<h2>List of Contacts:</h2>";
+
   if (!$added) {
-    echo " <a href='index.php?method=POST'>Add Contact 42</a>";
+    echo "<ul class='action-links'><li><a href='index.php?method=POST'> + Add Contact 42</a></li></ul>";
   }
+
+  echo "<pre>";
+ // print_r($result);
+  echo "</pre>";
+
+  // Create table.
+  echo "<table>
+          <thead>
+              <tr>
+                <th>Number</th>
+                <th>First Name</th>
+                <th>Last Name</th>
+                <th>Phone</th>
+                <th>Email</th>
+                <th>Actions</th>
+              </tr>
+          </thead>
+          <tbody>
+            ";
+  echo $output;
+
+  echo "</tbody>
+      </table>";
 }
 
 /**
@@ -112,18 +158,23 @@ if (!isset($_GET['method']) && !isset($_GET['id'])) {
  * Example: http://localhost/example/index.php?method=DELETE&id=42
  */
 if ($_GET['method'] === 'DELETE') {
-  $id = isset($_GET['id']) ? $_GET['id'] : '42';
+  $id = isset($_GET['id']) ? $_GET['id'] : null;
   $key = get_resource_key_field($service, $apiConfig, 'Contact');
 
   // Make the DELETE request.
-  $result = $service->Contact->delete($id, $key);
+  if ($id === '42') {
+    $result = $service->Contact->delete($id, $key);
 
-  echo "<a href='index.php'>Return to List</a>";
-  echo "<h2>Contact DELETE number $id:</h2>";
-  echo "<h2>Contact DELETE Result:</h2>";
-  echo "<pre>";
-  print_r($result);
-  echo "</pre>";
+    echo "<a href='index.php'>Return to List</a>";
+    echo "<h2>Contact DELETE number $id:</h2>";
+    echo "<h2>Contact DELETE Result:</h2>";
+    echo "<pre>";
+    print_r($result);
+    echo "</pre>";
+  } else {
+    echo "<a href='index.php'>Return to List</a>";
+    echo "<h2>Only Contact 42 can be deleted in this example.</h2>";
+  }
 }
 
 /**
@@ -141,7 +192,9 @@ if ($_GET['method'] === 'GET') {
   $result = $service->Contact->get($id, $key);
 
   echo "<a href='index.php'>Return to List</a>";
-  echo " <a href='index.php?method=DELETE&id=" . $result->data[$key] . "'>Delete</a>";
+  if ($value[$key] === '42') {
+    echo " <a href='index.php?method=DELETE&id=" . $result->data[$key] . "'>Delete</a>";
+  }
   echo " <a href='index.php?method=PATCH&id=" . $result->data[$key] . "'>Add a Comment</a>";
   echo "<h2>Contact GET number $id:</h2>";
   echo "<h2>Contact GET Result:</h2>";
@@ -262,7 +315,7 @@ if ($_GET['method'] === 'PATCH') {
   $contact = $service->Contact->get($id, $key);
 
   echo "<a href='index.php'>Return to List</a>";
-  echo '<h2>Contact Pre-PATCH Object:</h2><pre>';
+  echo '<h2>Contact Pre-PATCH Object:</h2>';
   echo "<pre>";
   print_r($contact);
   echo '</pre>';
@@ -279,7 +332,7 @@ if ($_GET['method'] === 'PATCH') {
     }
   }';
 
-  echo '<h2>Contact JSON-Patch Object:</h2><pre>';
+  echo '<h2>Contact JSON-Patch Object:</h2>';
   echo "<pre>";
   print_r($patch);
   echo '</pre>';
@@ -287,7 +340,7 @@ if ($_GET['method'] === 'PATCH') {
   // Use JSON-Patch library to apply the patch to our contact.
   $patched = JsonPatch::patch(json_encode($contact->data), $patch);
 
-  echo '<h2>Contact PATCHed Object:</h2><pre>';
+  echo '<h2>Contact PATCHed Object:</h2>';
   echo "<pre>";
   print_r($patched);
   echo '</pre>';
@@ -331,10 +384,15 @@ if ($_GET['method'] === 'PATCH') {
   echo '</pre>';
 }
 
+
+echo "    </div>
+        </body>
+      </html>";
+
 // We're not done yet. Remember to update the cached access token.
 // Remember to replace $_SESSION with a real database or memcached.
 if ($client->getAccessToken()) {
-  $_SESSION['token'] = $client->getAccessToken();
+  $_SESSION['access_token'] = $client->getAccessToken();
 }
 
 // Helper functoin to discover to key field for a resource from the JSON-Schema.
